refactor (backend): update imports accordingly after refactoring db_engine
